move docker client into imagerunner
