+++ internal/cmd/docker/cmd.go
@@ -10,7 +10,7 @@ import (
 )
 
 var (
-	registryClient        http.DockerRegistry
+	imageRunnerService    http.ImageRunner
 	registryClientTimeout = 1 * time.Minute
 )
 
@@ -34,7 +34,7 @@ func Command(preRun func(cmd *cobra.Command, args []string)) *cobra.Command {
 
 			creds := reg.Credentials()
 			url := reg.APIBaseURL()
-			registryClient = http.NewDockerRegistry(url, creds.Username, creds.AccessKey, registryClientTimeout)
+			imageRunnerService = http.NewImageRunner(url, creds, registryClientTimeout)
 
 			return nil
 		},
